# Food_Recipe_Swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Food-Recipe-API",
		"title": "Favorite-Food-Recipe-API"
	},
	"host": "localhost:8080",
	"basePath": "/",
	"tags": [
		{
			"name": "food-recipe-controller",
			"description": "Food Recipe Controller"
		}
	],
	"paths": {
		"/food-recipe-api/v1/checkRecipe/recipeName/{recipeName}": {
			"get": {
				"tags": [
					"food-recipe-controller"
				],
				"summary": "checkRecipeCategoryByRecipeName",
				"operationId": "checkRecipeCategoryByRecipeNameUsingGET",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "recipeName",
						"in": "path",
						"description": "recipeName",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/food-recipe-api/v1/createRecipe": {
			"post": {
				"tags": [
					"food-recipe-controller"
				],
				"summary": "createRecipes",
				"operationId": "createRecipesUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "recipes",
						"description": "recipes",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RecipesRequestDto"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/RecipesResponseDto"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/food-recipe-api/v1/deleteRecipe/{recipeId}": {
			"delete": {
				"tags": [
					"food-recipe-controller"
				],
				"summary": "deleteRecipe",
				"operationId": "deleteRecipeUsingDELETE",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "recipeId",
						"in": "path",
						"description": "recipeId",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			}
		},
		"/food-recipe-api/v1/getAllRecipe": {
			"get": {
				"tags": [
					"food-recipe-controller"
				],
				"summary": "getAllRecipes",
				"operationId": "getAllRecipesUsingGET",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/RecipesResponseDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/food-recipe-api/v1/getAllRecipe/category/{recipeCategory}": {
			"get": {
				"tags": [
					"food-recipe-controller"
				],
				"summary": "getAllRecipesByCategory",
				"operationId": "getAllRecipesByCategoryUsingGET",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "recipeCategory",
						"in": "path",
						"description": "recipeCategory",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/RecipesResponseDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/food-recipe-api/v1/getAllRecipe/ingredient/{ingredient}": {
			"get": {
				"tags": [
					"food-recipe-controller"
				],
				"summary": "getAllRecipesByIngredient",
				"operationId": "getAllRecipesByIngredientUsingGET",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "ingredient",
						"in": "path",
						"description": "ingredient",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/RecipesResponseDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/food-recipe-api/v1/getAllRecipe/ingredient/{ingredient}/instruction/{instruction}": {
			"get": {
				"tags": [
					"food-recipe-controller"
				],
				"summary": "getRecipeByInstructionNotIngredient",
				"operationId": "getRecipeByInstructionNotIngredientUsingGET",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "ingredient",
						"in": "path",
						"description": "ingredient",
						"required": true,
						"type": "string"
					},
					{
						"name": "instruction",
						"in": "path",
						"description": "instruction",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/RecipesResponseDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/food-recipe-api/v1/getAllRecipe/numberOfServes/{numberOfServes}/ingredient/{ingredient}": {
			"get": {
				"tags": [
					"food-recipe-controller"
				],
				"summary": "getRecipeByIngredientAndServes",
				"operationId": "getRecipeByIngredientAndServesUsingGET",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "numberOfServes",
						"in": "path",
						"description": "numberOfServes",
						"required": true,
						"type": "integer",
						"format": "int32"
					},
					{
						"name": "ingredient",
						"in": "path",
						"description": "ingredient",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/RecipesResponseDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/food-recipe-api/v1/updateRecipe": {
			"put": {
				"tags": [
					"food-recipe-controller"
				],
				"summary": "updateRecipe",
				"operationId": "updateRecipeUsingPUT",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "recipesUpdateRequestDto",
						"description": "recipesUpdateRequestDto",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RecipesUpdateRequestDto"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/RecipesResponseDto"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"RecipesRequestDto": {
			"type": "object",
			"properties": {
				"ingredients": {
					"type": "string"
				},
				"instructions": {
					"type": "string"
				},
				"noOfServings": {
					"type": "integer",
					"format": "int32"
				},
				"recipeCategory": {
					"type": "string"
				},
				"recipeName": {
					"type": "string"
				}
			},
			"title": "RecipesRequestDto"
		},
		"RecipesResponseDto": {
			"type": "object",
			"properties": {
				"ingredients": {
					"type": "string"
				},
				"instructions": {
					"type": "string"
				},
				"noOfServings": {
					"type": "integer",
					"format": "int32"
				},
				"recipeCategory": {
					"type": "string"
				},
				"recipeId": {
					"type": "integer",
					"format": "int32"
				},
				"recipeName": {
					"type": "string"
				}
			},
			"title": "RecipesResponseDto"
		},
		"RecipesUpdateRequestDto": {
			"type": "object",
			"properties": {
				"ingredients": {
					"type": "string"
				},
				"instructions": {
					"type": "string"
				},
				"noOfServings": {
					"type": "integer",
					"format": "int32"
				},
				"recipeCategory": {
					"type": "string"
				},
				"recipeId": {
					"type": "integer",
					"format": "int32"
				},
				"recipeName": {
					"type": "string"
				}
			},
			"title": "RecipesUpdateRequestDto"
		}
	}
}